Refatorando aplicação para deixar de usar API reativa que deixa a aplicação mais complexa

+++ src/main/resources/scripts/ddl/002_refresh_token.sql
@@ -1,10 +1,10 @@
 CREATE TABLE refresh_token
 (
-    id         SERIAL                      NOT NULL,
-    token      VARCHAR(36)                 NOT NULL,
-    user_id BIGINT                      NOT NULL,
-    creationDateTime  TIMESTAMP WITHOUT TIME ZONE NOT NULL,
-    expirationDateTime  TIMESTAMP WITHOUT TIME ZONE NOT NULL,
+    id                 BIGSERIAL                   NOT NULL,
+    token              VARCHAR(36)                 NOT NULL,
+    user_id            BIGINT                      NOT NULL,
+    creationDateTime   TIMESTAMP WITHOUT TIME ZONE NOT NULL,
+    expirationDateTime TIMESTAMP WITHOUT TIME ZONE NOT NULL,
     CONSTRAINT pk_refresh_token PRIMARY KEY (id),
     CONSTRAINT fk_user_refresh_token FOREIGN KEY (user_id) REFERENCES "user" (id)
 );
